Split LXC/rkt/Mesos support into separate files
Docker is a big chunk of the code itself so will be done
in a separate commit or two
