+++ userspace/libsinsp/container.cpp
@@ -25,6 +25,10 @@ limitations under the License.
 #include "cri.pb.h"
 #include "cri.grpc.pb.h"
 
+#include "container_rkt.h"
+#include "container_lxc.h"
+#include "container_mesos.h"
+
 #define CONTAINER_CPP
 typedef runtime::v1alpha2::RuntimeService::Stub RuntimeService_Stub;
 #endif
@@ -36,96 +40,6 @@ typedef runtime::v1alpha2::RuntimeService::Stub RuntimeService_Stub;
 #ifdef CYGWING_AGENT
 #include "dragent_win_hal_public.h"
 #endif
-
-const sinsp_container_info::container_mount_info *sinsp_container_info::mount_by_idx(uint32_t idx) const
-{
-	if (idx >= m_mounts.size())
-	{
-		return NULL;
-	}
-
-	return &(m_mounts[idx]);
-}
-
-const sinsp_container_info::container_mount_info *sinsp_container_info::mount_by_source(std::string &source) const
-{
-	// note: linear search
-	for (auto &mntinfo :m_mounts)
-	{
-		if(sinsp_utils::glob_match(source.c_str(), mntinfo.m_source.c_str()))
-		{
-			return &mntinfo;
-		}
-	}
-
-	return NULL;
-}
-
-const sinsp_container_info::container_mount_info *sinsp_container_info::mount_by_dest(std::string &dest) const
-{
-	// note: linear search
-	for (auto &mntinfo :m_mounts)
-	{
-		if(sinsp_utils::glob_match(dest.c_str(), mntinfo.m_dest.c_str()))
-		{
-			return &mntinfo;
-		}
-	}
-
-	return NULL;
-}
-
-std::string sinsp_container_info::normalize_healthcheck_arg(const std::string &arg)
-{
-	std::string ret = arg;
-
-	if(ret.empty())
-	{
-		return ret;
-	}
-
-	// Remove pairs of leading/trailing " or ' chars, if present
-	while(ret.front() == '"' || ret.front() == '\'')
-	{
-		if(ret.back() == ret.front())
-		{
-			ret.pop_back();
-			ret.erase(0, 1);
-		}
-	}
-
-	return ret;
-}
-
-void sinsp_container_info::parse_healthcheck(const Json::Value &healthcheck_obj)
-{
-	if(!healthcheck_obj.isNull())
-	{
-		const Json::Value &test_obj = healthcheck_obj["Test"];
-
-		if(!test_obj.isNull() && test_obj.isArray() && test_obj.size() >= 2)
-		{
-			if(test_obj[0].asString() == "CMD")
-			{
-				m_has_healthcheck = true;
-				m_healthcheck_exe = normalize_healthcheck_arg(test_obj[1].asString());
-
-				for(uint32_t i = 2; i < test_obj.size(); i++)
-				{
-					m_healthcheck_args.push_back(normalize_healthcheck_arg(test_obj[i].asString()));
-				}
-			}
-			else if(test_obj[0].asString() == "CMD-SHELL")
-			{
-				m_has_healthcheck = true;
-				m_healthcheck_exe = "/bin/sh";
-				m_healthcheck_args.push_back("-c");
-				m_healthcheck_args.push_back(test_obj[1].asString());
-			}
-		}
-	}
-}
-
 
 #if !defined(CYGWING_AGENT) && defined(HAS_CAPTURE)
 CURLM *sinsp_container_engine_docker::m_curlm = NULL;
@@ -804,472 +718,6 @@ sinsp_docker_response sinsp_container_engine_docker::get_docker(sinsp_container_
 #endif
 }
 
-bool sinsp_container_engine_lxc::resolve(sinsp_container_manager* manager, sinsp_threadinfo* tinfo, bool query_os_for_missing_info)
-{
-	sinsp_container_info container_info;
-	bool matches = false;
-
-	for(auto it = tinfo->m_cgroups.begin(); it != tinfo->m_cgroups.end(); ++it)
-	{
-		string cgroup = it->second;
-		size_t pos;
-
-		//
-		// Non-systemd LXC
-		//
-		pos = cgroup.find("/lxc/");
-		if(pos != string::npos)
-		{
-			auto id_start = pos + sizeof("/lxc/") - 1;
-			auto id_end = cgroup.find('/', id_start);
-			container_info.m_type = CT_LXC;
-			container_info.m_id = cgroup.substr(id_start, id_end - id_start);
-			matches = true;
-			break;
-		}
-	}
-
-	if (!matches)
-		return false;
-
-	tinfo->m_container_id = container_info.m_id;
-	if (!manager->container_exists(container_info.m_id))
-	{
-		container_info.m_name = container_info.m_id;
-		manager->add_container(container_info, tinfo);
-		manager->notify_new_container(container_info);
-	}
-	return true;
-}
-
-bool sinsp_container_engine_libvirt_lxc::match(sinsp_threadinfo* tinfo, sinsp_container_info* container_info)
-{
-	for(auto it = tinfo->m_cgroups.begin(); it != tinfo->m_cgroups.end(); ++it)
-	{
-		string cgroup = it->second;
-		size_t pos;
-
-		//
-		// Non-systemd libvirt-lxc
-		//
-		pos = cgroup.find(".libvirt-lxc");
-		if(pos != string::npos &&
-			pos == cgroup.length() - sizeof(".libvirt-lxc") + 1)
-		{
-			size_t pos2 = cgroup.find_last_of("/");
-			if(pos2 != string::npos)
-			{
-				container_info->m_type = CT_LIBVIRT_LXC;
-				container_info->m_id = cgroup.substr(pos2 + 1, pos - pos2 - 1);
-				return true;
-			}
-		}
-
-		//
-		// systemd libvirt-lxc
-		//
-		pos = cgroup.find("-lxc\\x2");
-		if(pos != string::npos)
-		{
-			size_t pos2 = cgroup.find(".scope");
-			if(pos2 != string::npos &&
-				pos2 == cgroup.length() - sizeof(".scope") + 1)
-			{
-				container_info->m_type = CT_LIBVIRT_LXC;
-				container_info->m_id = cgroup.substr(pos + sizeof("-lxc\\x2"), pos2 - pos - sizeof("-lxc\\x2"));
-				return true;
-			}
-		}
-
-		//
-		// Legacy libvirt-lxc
-		//
-		pos = cgroup.find("/libvirt/lxc/");
-		if(pos != string::npos)
-		{
-			container_info->m_type = CT_LIBVIRT_LXC;
-			container_info->m_id = cgroup.substr(pos + sizeof("/libvirt/lxc/") - 1);
-			return true;
-		}
-	}
-	return false;
-}
-
-bool sinsp_container_engine_libvirt_lxc::resolve(sinsp_container_manager* manager, sinsp_threadinfo* tinfo, bool query_os_for_missing_info)
-{
-	sinsp_container_info container_info;
-
-	if (!match(tinfo, &container_info))
-		return false;
-	
-	tinfo->m_container_id = container_info.m_id;
-	if (!manager->container_exists(container_info.m_id))
-	{
-		container_info.m_name = container_info.m_id;
-		manager->add_container(container_info, tinfo);
-		manager->notify_new_container(container_info);
-	}
-	return true;
-}
-
-
-bool sinsp_container_engine_mesos::match(sinsp_threadinfo* tinfo, sinsp_container_info* container_info)
-{
-	for(auto it = tinfo->m_cgroups.begin(); it != tinfo->m_cgroups.end(); ++it)
-	{
-		string cgroup = it->second;
-		size_t pos;
-
-		pos = cgroup.find("/mesos/");
-		if(pos != string::npos)
-		{
-			// It should match `/mesos/a9f41620-b165-4d24-abe0-af0af92e7b20`
-			auto id = cgroup.substr(pos + sizeof("/mesos/") - 1);
-			if(id.size() == 36 && id.find_first_not_of("0123456789abcdefABCDEF-") == string::npos)
-			{
-				container_info->m_type = CT_MESOS;
-				container_info->m_id = move(id);
-				// Consider a mesos container valid only if we find the mesos_task_id
-				// this will exclude from the container itself the mesos-executor
-				// but makes sure that we have task_id parsed properly. Otherwise what happens
-				// is that we'll create a mesos container struct without a mesos_task_id
-				// and for all other processes we'll use it
-				return set_mesos_task_id(container_info, tinfo);
-			}
-		}
-	}
-	return false;
-}
-
-bool sinsp_container_engine_mesos::resolve(sinsp_container_manager* manager, sinsp_threadinfo* tinfo, bool query_os_for_missing_info)
-{
-	sinsp_container_info container_info;
-
-	if (!match(tinfo, &container_info))
-		return false;
-	
-	tinfo->m_container_id = container_info.m_id;
-	if (!manager->container_exists(container_info.m_id))
-	{
-		container_info.m_name = container_info.m_id;
-		manager->add_container(container_info, tinfo);
-		manager->notify_new_container(container_info);
-	}
-	return true;
-}
-
-string sinsp_container_engine_mesos::get_env_mesos_task_id(sinsp_threadinfo* tinfo)
-{
-	string mtid;
-
-	sinsp_threadinfo::visitor_func_t visitor = [&mtid] (sinsp_threadinfo *ptinfo)
-	{
-		// Mesos task ID detection is not a straightforward task;
-		// this list may have to be extended.
-		mtid = ptinfo->get_env("MESOS_TASK_ID"); // Marathon
-		if(!mtid.empty()) { return false; }
-		mtid = ptinfo->get_env("mesos_task_id"); // Chronos
-		if(!mtid.empty()) { return false; }
-		mtid = ptinfo->get_env("MESOS_EXECUTOR_ID"); // others
-		if(!mtid.empty()) { return false; }
-
-		return true;
-	};
-
-	// Try the current thread first. visitor returns true if mtid
-	// was not filled in. In this case we should traverse the
-	// parents.
-	if(tinfo && visitor(tinfo))
-	{
-		tinfo->traverse_parent_state(visitor);
-	}
-
-	return mtid;
-}
-
-bool sinsp_container_engine_mesos::set_mesos_task_id(sinsp_container_info* container, sinsp_threadinfo* tinfo)
-{
-	ASSERT(container);
-	ASSERT(tinfo);
-
-	// there are applications that do not share their environment in /proc/[PID]/environ
-	// since we need MESOS_TASK_ID environment variable to discover Mesos containers,
-	// there is a workaround for such cases:
-	// - for docker containers, we discover it directly from container, through Remote API
-	//   (see sinsp_container_manager::parse_docker() for details)
-	// - for mesos native containers, parent process has the MESOS_TASK_ID (or equivalent, see
-	//   get_env_mesos_task_id(sinsp_threadinfo*) implementation) environment variable, so we
-	//   peek into the parent process environment to discover it
-
-	if(container && tinfo)
-	{
-		string& mtid = container->m_mesos_task_id;
-		if(mtid.empty())
-		{
-			mtid = get_env_mesos_task_id(tinfo);
-
-			// Ensure that the mesos task id vaguely looks
-			// like a real id. We assume it must be at
-			// least 3 characters and contain a dot or underscore
-			if(!mtid.empty() && mtid.length()>=3 &&
-			   (mtid.find_first_of("._") != std::string::npos))
-			{
-				g_logger.log("Mesos native container: [" + container->m_id + "], Mesos task ID: " + mtid, sinsp_logger::SEV_DEBUG);
-				return true;
-			}
-			else
-			{
-				g_logger.log("Mesos container [" + container->m_id + "],"
-					     "thread [" + std::to_string(tinfo->m_tid) +
-					     "], has likely malformed mesos task id [" + mtid + "], ignoring", sinsp_logger::SEV_DEBUG);
-			}
-		}
-	}
-	return false;
-}
-
-bool sinsp_container_engine_rkt::match(sinsp_container_manager* manager, sinsp_threadinfo* tinfo, sinsp_container_info* container_info, string& rkt_podid, string& rkt_appname, bool query_os_for_missing_info)
-{
-	for(auto it = tinfo->m_cgroups.begin(); it != tinfo->m_cgroups.end(); ++it)
-	{
-		string cgroup = it->second;
-
-		static const string COREOS_PODID_VAR = "container_uuid=";
-		static const string SYSTEMD_UUID_ARG = "--uuid=";
-		static const string SERVICE_SUFFIX = ".service";
-		if(cgroup.rfind(SERVICE_SUFFIX) == cgroup.size() - SERVICE_SUFFIX.size())
-		{
-			// check if there is a parent with pod uuid var
-			sinsp_threadinfo::visitor_func_t visitor = [&](sinsp_threadinfo* ptinfo)
-			{
-				for(const auto& env_var : ptinfo->get_env())
-				{
-					auto container_uuid_pos = env_var.find(COREOS_PODID_VAR);
-					if(container_uuid_pos == 0)
-					{
-						rkt_podid = env_var.substr(COREOS_PODID_VAR.size());
-						return false;
-					}
-				}
-				for(const auto& arg : ptinfo->m_args)
-				{
-					if(arg.find(SYSTEMD_UUID_ARG) != string::npos)
-					{
-						rkt_podid = arg.substr(SYSTEMD_UUID_ARG.size());
-						return false;
-					}
-				}
-				return true;
-			};
-			tinfo->traverse_parent_state(visitor);
-
-			if(!rkt_podid.empty())
-			{
-				auto last_slash = cgroup.find_last_of("/");
-				rkt_appname = cgroup.substr(last_slash + 1, cgroup.size() - last_slash - SERVICE_SUFFIX.size() - 1);
-
-				char image_manifest_path[SCAP_MAX_PATH_SIZE];
-				snprintf(image_manifest_path, sizeof(image_manifest_path), "%s/var/lib/rkt/pods/run/%s/appsinfo/%s/manifest", scap_get_host_root(), rkt_podid.c_str(), rkt_appname.c_str());
-
-				// First lookup if the container exists in our table, otherwise only if we are live check if it has
-				// an entry in /var/lib/rkt. In capture mode only the former will be used.
-				// In live mode former will be used only if we already hit that container
-				bool is_rkt_pod_id_valid = manager->container_exists(rkt_podid + ":" + rkt_appname); // if it's already on our table
-#ifdef HAS_CAPTURE
-				if(!is_rkt_pod_id_valid && query_os_for_missing_info)
-				{
-					is_rkt_pod_id_valid = (access(image_manifest_path, F_OK) == 0);
-				}
-#endif
-				if(is_rkt_pod_id_valid)
-				{
-					container_info->m_type = CT_RKT;
-					container_info->m_id = rkt_podid + ":" + rkt_appname;
-					container_info->m_name = rkt_appname;
-					return true;
-				}
-			}
-		}
-	}
-
-	// Try parsing from process root,
-	// Strings used to detect rkt stage1-cores pods
-	// TODO: detecting stage1-coreos rkt pods in this way is deprecated
-	// we can remove it in the future
-	static const string COREOS_PREFIX = "/opt/stage2/";
-	static const string COREOS_APP_SUFFIX = "/rootfs";
-	static const string COREOS_PODID_VAR = "container_uuid=";
-
-	auto prefix = tinfo->m_root.find(COREOS_PREFIX);
-	if(prefix == 0)
-	{
-		auto suffix = tinfo->m_root.find(COREOS_APP_SUFFIX, prefix);
-		if(suffix != string::npos)
-		{
-			bool valid_id = false;
-			rkt_appname = tinfo->m_root.substr(prefix + COREOS_PREFIX.size(), suffix - prefix - COREOS_PREFIX.size());
-			// It is a rkt pod with stage1-coreos
-
-			sinsp_threadinfo::visitor_func_t visitor = [&] (sinsp_threadinfo *ptinfo)
-			{
-				for(const auto& env_var : ptinfo->get_env())
-				{
-					auto container_uuid_pos = env_var.find(COREOS_PODID_VAR);
-					if(container_uuid_pos == 0)
-					{
-						rkt_podid = env_var.substr(COREOS_PODID_VAR.size());
-						container_info->m_type = CT_RKT;
-						container_info->m_id = rkt_podid + ":" + rkt_appname;
-						container_info->m_name = rkt_appname;
-						valid_id = true;
-						return false;
-					}
-				}
-				return true;
-			};
-
-			// Try the current thread first. visitor returns true if no coreos pid
-			// info was found. In this case we traverse the parents.
-			if (visitor(tinfo))
-			{
-				tinfo->traverse_parent_state(visitor);
-			}
-			return valid_id;
-		}
-	}
-	else
-	{
-		// String used to detect stage1-fly pods
-		static const string FLY_PREFIX = "/var/lib/rkt/pods/run/";
-		static const string FLY_PODID_SUFFIX = "/stage1/rootfs/opt/stage2/";
-		static const string FLY_APP_SUFFIX = "/rootfs";
-
-		auto prefix = tinfo->m_root.find(FLY_PREFIX);
-		if(prefix == 0)
-		{
-			auto podid_suffix = tinfo->m_root.find(FLY_PODID_SUFFIX, prefix+FLY_PREFIX.size());
-			if(podid_suffix != string::npos)
-			{
-				rkt_podid = tinfo->m_root.substr(prefix + FLY_PREFIX.size(), podid_suffix - prefix - FLY_PREFIX.size());
-				auto appname_suffix = tinfo->m_root.find(FLY_APP_SUFFIX, podid_suffix+FLY_PODID_SUFFIX.size());
-				if(appname_suffix != string::npos)
-				{
-					rkt_appname = tinfo->m_root.substr(podid_suffix + FLY_PODID_SUFFIX.size(),
-													   appname_suffix-podid_suffix-FLY_PODID_SUFFIX.size());
-					container_info->m_type = CT_RKT;
-					container_info->m_id = rkt_podid + ":" + rkt_appname;
-					container_info->m_name = rkt_appname;
-					return true;
-				}
-			}
-		}
-	}
-	return false;
-}
-
-bool sinsp_container_engine_rkt::resolve(sinsp_container_manager* manager, sinsp_threadinfo* tinfo, bool query_os_for_missing_info)
-{
-	sinsp_container_info container_info;
-	string rkt_podid, rkt_appname;
-
-	if (!match(manager, tinfo, &container_info, rkt_podid, rkt_appname, query_os_for_missing_info))
-	{
-		return false;
-	}
-
-	tinfo->m_container_id = container_info.m_id;
-	if (!query_os_for_missing_info || manager->container_exists(container_info.m_id))
-	{
-		return true;
-	}
-
-#ifndef _WIN32
-	bool have_rkt = parse_rkt(&container_info, rkt_podid, rkt_appname);
-#else
-	bool have_rkt = true;
-#endif
-
-	if (have_rkt)
-	{
-		manager->add_container(container_info, tinfo);
-		manager->notify_new_container(container_info);
-		return true;
-	}
-	else
-	{
-		return false;
-	}
-}
-
-bool sinsp_container_engine_rkt::parse_rkt(sinsp_container_info *container, const string &podid, const string &appname)
-{
-	bool ret = false;
-	Json::Reader reader;
-	Json::Value jroot;
-
-	char image_manifest_path[SCAP_MAX_PATH_SIZE];
-	snprintf(image_manifest_path, sizeof(image_manifest_path), "%s/var/lib/rkt/pods/run/%s/appsinfo/%s/manifest", scap_get_host_root(), podid.c_str(), appname.c_str());
-	ifstream image_manifest(image_manifest_path);
-	if(reader.parse(image_manifest, jroot))
-	{
-		container->m_image = jroot["name"].asString();
-		for(const auto& label_entry : jroot["labels"])
-		{
-			container->m_labels.emplace(label_entry["name"].asString(), label_entry["value"].asString());
-		}
-		auto version_label_it = container->m_labels.find("version");
-		if(version_label_it != container->m_labels.end())
-		{
-			container->m_image += ":" + version_label_it->second;
-		}
-		ret = true;
-	}
-
-	char net_info_path[SCAP_MAX_PATH_SIZE];
-	snprintf(net_info_path, sizeof(net_info_path), "%s/var/lib/rkt/pods/run/%s/net-info.json", scap_get_host_root(), podid.c_str());
-	ifstream net_info(net_info_path);
-	if(reader.parse(net_info, jroot) && jroot.size() > 0)
-	{
-		const auto& first_net = jroot[0];
-		if(inet_pton(AF_INET, first_net["ip"].asCString(), &container->m_container_ip) == -1)
-		{
-			ASSERT(false);
-		}
-		container->m_container_ip = ntohl(container->m_container_ip);
-	}
-
-	char pod_manifest_path[SCAP_MAX_PATH_SIZE];
-	snprintf(pod_manifest_path, sizeof(pod_manifest_path), "%s/var/lib/rkt/pods/run/%s/pod", scap_get_host_root(), podid.c_str());
-	ifstream pod_manifest(pod_manifest_path);
-	unordered_map<string, uint32_t> image_ports;
-	if(reader.parse(pod_manifest, jroot) && jroot.size() > 0)
-	{
-		for(const auto& japp : jroot["apps"])
-		{
-			if (japp["name"].asString() == appname)
-			{
-				for(const auto& image_port : japp["app"]["ports"])
-				{
-					image_ports[image_port["name"].asString()] = image_port["port"].asUInt();
-				}
-				break;
-			}
-		}
-		for(const auto& jport : jroot["ports"])
-		{
-			auto host_port = jport["hostPort"].asUInt();
-			auto container_port_it = image_ports.find(jport["name"].asString());
-			if(host_port > 0 && container_port_it != image_ports.end())
-			{
-				sinsp_container_info::container_port_mapping port_mapping;
-				port_mapping.m_host_port = host_port;
-				port_mapping.m_container_port = container_port_it->second;
-				container->m_port_mappings.emplace_back(move(port_mapping));
-			}
-		}
-	}
-	return ret;
-}
 #endif
 
 sinsp_container_manager::sinsp_container_manager(sinsp* inspector) :
