+++ userspace/libsinsp/container.h
@@ -21,6 +21,8 @@ limitations under the License.
 
 #include <functional>
 
+#include "container_info.h"
+
 #if !defined(_WIN32) && !defined(CYGWING_AGENT) && defined(HAS_CAPTURE)
 #include <curl/curl.h>
 #include <curl/easy.h>
@@ -34,152 +36,11 @@ class RuntimeService_Stub;
 #endif
 #endif
 
-enum sinsp_container_type
-{
-	CT_DOCKER = 0,
-	CT_LXC = 1,
-	CT_LIBVIRT_LXC = 2,
-	CT_MESOS = 3,
-	CT_RKT = 4,
-	CT_CUSTOM = 5
-};
-
 enum sinsp_docker_response
 {
 	RESP_OK = 0,
 	RESP_BAD_REQUEST = 1,
 	RESP_ERROR = 2
-};
-
-class sinsp_container_info
-{
-public:
-
-	class container_port_mapping
-	{
-	public:
-		container_port_mapping():
-			m_host_ip(0),
-			m_host_port(0),
-			m_container_port(0)
-		{
-		}
-		uint32_t m_host_ip;
-		uint16_t m_host_port;
-		uint16_t m_container_port;
-	};
-
-	class container_mount_info
-	{
-	public:
-		container_mount_info():
-			m_source(""),
-			m_dest(""),
-			m_mode(""),
-			m_rdwr(false),
-			m_propagation("")
-		{
-		}
-
-		container_mount_info(const std::string&& source, const std::string&& dest,
-				     const std::string&& mode, const bool rw,
-				     const std::string&& propagation) :
-			m_source(source), m_dest(dest), m_mode(mode), m_rdwr(rw), m_propagation(propagation)
-		{
-		}
-
-		container_mount_info(const Json::Value &source, const Json::Value &dest,
-				     const Json::Value &mode, const Json::Value &rw,
-				     const Json::Value &propagation)
-		{
-			get_string_value(source, m_source);
-			get_string_value(dest, m_dest);
-			get_string_value(mode, m_mode);
-			get_string_value(propagation, m_propagation);
-
-			if(!rw.isNull() && rw.isBool())
-			{
-				m_rdwr = rw.asBool();
-			}
-		}
-
-		std::string to_string() const
-		{
-			return m_source + ":" +
-				m_dest + ":" +
-				m_mode + ":" +
-				(m_rdwr ? "true" : "false") + ":" +
-				m_propagation;
-		}
-
-		inline void get_string_value(const Json::Value &val, std::string &result)
-		{
-			if(!val.isNull() && val.isString())
-			{
-				result = val.asString();
-			}
-		}
-
-		std::string m_source;
-		std::string m_dest;
-		std::string m_mode;
-		bool m_rdwr;
-		std::string m_propagation;
-	};
-
-	sinsp_container_info():
-		m_container_ip(0),
-		m_privileged(false),
-		m_memory_limit(0),
-		m_swap_limit(0),
-		m_cpu_shares(1024),
-		m_cpu_quota(0),
-		m_cpu_period(100000),
-		m_has_healthcheck(false),
-		m_healthcheck_exe("")
-#ifdef HAS_ANALYZER
-		,m_metadata_deadline(0)
-#endif
-	{
-	}
-
-	std::string normalize_healthcheck_arg(const std::string &arg);
-	void parse_healthcheck(const Json::Value &config_obj);
-
-	const vector<string>& get_env() const { return m_env; }
-
-	const container_mount_info *mount_by_idx(uint32_t idx) const;
-	const container_mount_info *mount_by_source(std::string &source) const;
-	const container_mount_info *mount_by_dest(std::string &dest) const;
-
-	string m_id;
-	sinsp_container_type m_type;
-	string m_name;
-	string m_image;
-	string m_imageid;
-	string m_imagerepo;
-	string m_imagetag;
-	string m_imagedigest;
-	uint32_t m_container_ip;
-	bool m_privileged;
-	vector<container_mount_info> m_mounts;
-	vector<container_port_mapping> m_port_mappings;
-	map<string, string> m_labels;
-	vector<string> m_env;
-	string m_mesos_task_id;
-	int64_t m_memory_limit;
-	int64_t m_swap_limit;
-	int64_t m_cpu_shares;
-	int64_t m_cpu_quota;
-	int64_t m_cpu_period;
-	Json::Value m_healthcheck_obj;
-	bool m_has_healthcheck;
-	std::string m_healthcheck_exe;
-	std::vector<std::string> m_healthcheck_args;
-#ifdef HAS_ANALYZER
-	string m_sysdig_agent_conf;
-	uint64_t m_metadata_deadline;
-#endif
 };
 
 class sinsp_container_manager;
@@ -215,42 +76,6 @@ protected:
 	static unique_ptr<RuntimeService_Stub> m_containerd;
 #endif
 };
-
-#ifndef CYGWING_AGENT
-class sinsp_container_engine_lxc
-{
-public:
-	bool resolve(sinsp_container_manager* manager, sinsp_threadinfo* tinfo, bool query_os_for_missing_info);
-};
-
-class sinsp_container_engine_libvirt_lxc
-{
-public:
-	bool resolve(sinsp_container_manager* manager, sinsp_threadinfo* tinfo, bool query_os_for_missing_info);
-protected:
-	bool match(sinsp_threadinfo* tinfo, sinsp_container_info* container_info);
-};
-
-class sinsp_container_engine_mesos
-{
-public:
-	bool resolve(sinsp_container_manager* manager, sinsp_threadinfo* tinfo, bool query_os_for_missing_info);
-	static bool set_mesos_task_id(sinsp_container_info* container, sinsp_threadinfo* tinfo);
-protected:
-	bool match(sinsp_threadinfo* tinfo, sinsp_container_info* container_info);
-	static string get_env_mesos_task_id(sinsp_threadinfo* tinfo);
-};
-
-class sinsp_container_engine_rkt
-{
-public:
-	bool resolve(sinsp_container_manager* manager, sinsp_threadinfo* tinfo, bool query_os_for_missing_info);
-protected:
-	bool match(sinsp_container_manager* manager, sinsp_threadinfo* tinfo, sinsp_container_info* container_info, string& rkt_podid, string& rkt_appname, bool query_os_for_missing_info);
-	bool parse_rkt(sinsp_container_info* container, const string& podid, const string& appname);
-};
-
-#endif
 
 class sinsp_container_manager
 {
