+++ userspace/libsinsp/container_rkt.h
@@ -1,0 +1,36 @@
+/*
+Copyright (C) 2013-2018 Draios Inc dba Sysdig.
+
+This file is part of sysdig.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+
+*/
+
+#pragma once
+
+#include <string>
+
+class sinsp_container_manager;
+class sinsp_container_info;
+class sinsp_threadinfo;
+
+class sinsp_container_engine_rkt
+{
+public:
+	bool resolve(sinsp_container_manager* manager, sinsp_threadinfo* tinfo, bool query_os_for_missing_info);
+protected:
+	bool match(sinsp_container_manager* manager, sinsp_threadinfo* tinfo, sinsp_container_info* container_info, std::string& rkt_podid, std::string& rkt_appname, bool query_os_for_missing_info);
+	bool parse_rkt(sinsp_container_info* container, const std::string& podid, const std::string& appname);
+};
+
